added encoder calibration

+++ SteerByWireV2/Core/Inc/encoder.hpp
@@ -16,16 +16,16 @@ class Encoder{
 		MotorControl* motorControl;
 		TIM_HandleTypeDef* encoder1;
 		TIM_HandleTypeDef* encoder2;
-		uint32_t maxEncoderCount;
+		uint16_t maxEncoderCount;
 		float countsPerRev;
 
 	public:
-		Encoder(MotorControl* motorControl, TIM_HandleTypeDef* enc1, TIM_HandleTypeDef* enc2, uint32_t maxCount, float countsPerRevolution);
+		Encoder(MotorControl* motorControl, TIM_HandleTypeDef* enc1, TIM_HandleTypeDef* enc2, uint16_t maxCount, float countsPerRevolution);
 		uint32_t readEncoder(TIM_HandleTypeDef* enc);
 		void resetEncoderCount(TIM_HandleTypeDef* enc);
 		void calibrateEncoder(TIM_HandleTypeDef* enc, uint16_t motorCurrent);
-		float calculateAngle(uint32_t count);
-		float encoderCompare();
+		int16_t calculateAngle(uint16_t count);
+		int16_t encoderCompare();
 };
